Add files via upload
The latest version of the algorithm to generate a neighborhood around an arbitrary point in the plane. In this version we have corrected some errors present when calculating the dynamics of the particle if the rotational symmetry of the lattice was even (changes implemented in the file "Esferas_Duras_Trayectoria_Recta.jl"). We have also modified the functions present in the file "Algoritmo_Dinamica.jl" for the particular case of calculating the free path length distribution, generating a new file called "Algoritmo_Dinamica_FFD.jl"