Update algorithm to get the main cluster
Files with the update of the new algorithm to obtain the main cluster of tiles of the quasiperiodic array around an arbitrary point in the plane. The previous version iterated an algorithm based on an upper bound for the area of ​​the Voronoi tiles associated with the centroids of the quasiperiodic tiles N number of times. The new algorithm uses the upper bound for the area of ​​the Voronoi tiles to identify the border tile closest to the center of the generated patch, then removes all tiles that are further from the center than the closest border tile.