Delete Esferas_Duras_Trayectoria_Recta.jl
This file corresponds to a routine to calculate a potential of hard spheres in a Lorentz gas considering only the obstacle associated with the Voronoi cell that contains the particle. A version considering first neighbors has been developed to generalize this, that is why this file is removed to add the two cases separately with different names.